+++ src/TestSuite/ComponentTestCase.php
@@ -24,7 +24,7 @@ abstract class ComponentTestCase extends TestCase
 {
     /**
      * Component instance
-     * @var \Cake\Controller\Component|\PHPUnit_Framework_MockObject_MockObject
+     * @var \Cake\Controller\Component|\PHPUnit\Framework\MockObject\MockObject
      */
     protected $Component;